+++ packages/zero-cache/src/services/change-streamer/change-streamer-http.ts
@@ -18,7 +18,7 @@ import {URLParams} from '../../types/url-params.ts';
 import {installWebSocketReceiver} from '../../types/websocket-handoff.ts';
 import {closeWithError, PROTOCOL_ERROR} from '../../types/ws.ts';
 import {HttpService} from '../http-service.ts';
-import type {SnapshotReservationMonitor} from './backup-monitor-interface.ts';
+import type {BackupMonitor} from './backup-monitor.ts';
 import {
   downstreamSchema,
   PROTOCOL_VERSION,
@@ -50,14 +50,14 @@ export class ChangeStreamerHttpServer extends HttpService {
   readonly #lc: LogContext;
   readonly #opts: Options;
   readonly #changeStreamer: ChangeStreamerService;
-  readonly #backupMonitor: SnapshotReservationMonitor | null;
+  readonly #backupMonitor: BackupMonitor | null;
 
   constructor(
     lc: LogContext,
     opts: Options,
     parent: Worker,
     changeStreamer: ChangeStreamerService,
-    backupMonitor: SnapshotReservationMonitor | null,
+    backupMonitor: BackupMonitor | null,
   ) {
     super('change-streamer-http-server', lc, opts, async fastify => {
       await fastify.register(websocket);
